auth in hooks
