+++ src/services/device/index.js
@@ -29,39 +29,25 @@ class Service {
     };
 
     let newUser = new user({
-      schulcloudId: null,
+      schulcloudId: data.schulcloudId,
       devices: [newDevice]
     });
 
     // Insert data into DB
-    return new Promise((resolve, reject) => {
-
-      // Resolve user only existing users can register in the Schul-Cloud app
-      Resolve.verifyUser(data.user_token)
-        .then(schulcloudId => {
-          newUser.schulcloudId = schulcloudId;
-          return user.findOne({ schulcloudId: schulcloudId })
-        })
-        .then(user => {
-          if (!user) {
-            var user = newUser;
-          } else {
-            const deviceExists = user.devices.some(device => {
-              if (device.token === newDevice.token) return true;
-            });
-            if (!deviceExists) {
-              user.devices.push(newDevice);
-            }
+    return user.findOne({ schulcloudId: data.schulcloudId })
+      .then(user => {
+        if (!user) {
+          var user = newUser;
+        } else {
+          const deviceExists = user.devices.some(device => {
+            if (device.token === newDevice.token) return true;
+          });
+          if (!deviceExists) {
+            user.devices.push(newDevice);
           }
-          return user.save();
-        })
-        .then(userWithNewDevice => {
-          resolve(userWithNewDevice);
-        })
-        .catch(err => {
-          reject(err);
-        })
-    });
+        }
+        return user.save();
+      })
   }
 }
